requirements.txt
# Core ADK dependency
google-adk

# Environment variables
python-dotenv

# Optional: Supabase for database
# supabase

# Optional: HTTP requests
# httpx

# Optional: Testing
# pytest
